requirements.txt
pyserial>=3.5
matplotlib>=3.5
numpy>=1.21
scipy>=1.7
pandas>=1.3
